# docs/openapi.yml
openapi: 3.0.3
info:
  title: BetterIDN API
  description: API documentation for BetterIDN application
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev server
paths:
...
  /posts:
    post:
      summary: Create a new post
      tags:
        - posts
      description: Creates a new post with title, content, and categories.
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        "201":
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePostSuccessResponse'
        "400":
          description: Bad request (validation error or invalid request body)
        "401":
          description: Unauthorized (user not logged in)
        "500":
          description: Internal server error

    get:
      summary: Get paginated list of posts
      tags:
        - posts
      description: Retrieves a paginated list of posts.
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of posts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPostsSuccessResponse'
        "400":
          description: Invalid pagination parameters
        "500":
          description: Internal server error
  /posts/{postId}:
    get:
      summary: Get a post by ID
      tags:
        - posts
      description: Retrieves a specific post by its ID.
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPostSuccessResponse'
        "400":
          description: Invalid post ID format
        "404":
          description: Post not found
        "500":
          description: Internal server error

    put:
      summary: Update a post
      tags:
        - posts
      description: Updates an existing post with new title, content, and categories.
      security:
        - CookieAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        "200":
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePostSuccessResponse'
        "400":
          description: Bad request (validation error or invalid request body)
        "401":
          description: Unauthorized (user not logged in)
        "403":
          description: Forbidden (user not authorized to update this post)
        "404":
          description: Post not found
        "500":
          description: Internal server error
...
components:
  schemas:
...
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
          example: "My First Post"
        content:
          type: string
          example: "This is the content of my first post."
        categories:
          type: array
          items:
            type: string
          example: ["technology", "golang"]
      required:
        - title
        - content
        - categories
    CreatePostSuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "post created successfully"
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: "4fa85f64-5717-4562-b3fc-2c963f66afa6"
    GetPostsSuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "posts retrieved successfully"
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
                example: "4fa85f64-5717-4562-b3fc-2c963f66afa6"
              title:
                type: string
                example: "My First Post"
              content:
                type: string
                example: "This is the content of my first post."
              categories:
                type: array
                items:
                  type: string
                example: ["technology", "golang"]
    GetPostSuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "post retrieved successfully"
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: "4fa85f64-5717-4562-b3fc-2c963f66afa6"
            title:
              type: string
              example: "My First Post"
            content:
              type: string
              example: "This is the content of my first post."
            categories:
              type: array
              items:
                type: string
              example: ["technology", "golang"]
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
          example: "Updated Title"
        content:
          type: string
          example: "This is the updated content."
        categories:
          type: array
          items:
            type: string
          example: ["updated-category"]
      required:
        - title
        - content
        - categories
    UpdatePostSuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "post updated successfully"
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: "4fa85f64-5717-4562-b3fc-2c963f66afa6"
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token